Removing interfaces and refactoring
