+++ athena-gcs/src/main/java/com/amazonaws/athena/connectors/gcs/GcsConstants.java
@@ -59,10 +59,6 @@ public class GcsConstants
      */
     public static final String GOOGLE_SERVICE_ACCOUNT_JSON_TEMP_FILE_LOCATION = "GOOGLE_APPLICATION_CREDENTIALS";
 
-    public static final String FUNCTION_ARN_REGEX = "arn:aws[a-zA-Z-]*?:lambda:[a-zA-Z0-9-]+:(\\d{12}):function:[a-zA-Z0-9-_]+";
-
-    public static final String CATALOG_NAME_ENV_OVERRIDE = "glue_catalog";
-
     public static final String CLASSIFICATION_GLUE_TABLE_PARAM = "classification";
 
     public static final String PARTITION_PATTERN_PATTERN = "partition.pattern";
