--- athena-gcs/src/main/java/com/amazonaws/athena/connectors/gcs/storage/StorageConstants.java
@@ -1,35 +1,0 @@
-/*-
- * #%L
- * athena-hive
- * %%
- * Copyright (C) 2019 - 2022 Amazon Web Services
- * %%
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * #L%
- */
-package com.amazonaws.athena.connectors.gcs.storage;
-
-public class StorageConstants
-{
-    private StorageConstants()
-    {
-    }
-
-    /**
-     * A key that stores JSON representation of {@link com.amazonaws.athena.connectors.gcs.storage.StorageSplit} as a string
-     */
-    public static final String STORAGE_SPLIT_JSON = "storage_split_json";
-
-    // split constants
-    public static final String TABLE_PARAM_OBJECT_NAME_LIST = "objectName";
-}
